tweak: adjust biome distribution, fix configured features, update project structure
