+++ src/main/java/de/prwh/cobaltmod/core/block/CMPlantBlock.java
@@ -1,16 +1,15 @@
-package de.prwh.cobaltmod.core.blocks;
+package de.prwh.cobaltmod.core.block;
 
 import de.prwh.cobaltmod.core.tag.CMBlockTags;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.SaplingBlock;
-import net.minecraft.block.sapling.SaplingGenerator;
+import net.minecraft.block.PlantBlock;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.BlockView;
 
-public class CMSaplingBlock extends SaplingBlock {
+public class CMPlantBlock extends PlantBlock {
 
-    public CMSaplingBlock(SaplingGenerator generator, Settings settings) {
-        super(generator, settings);
+    public CMPlantBlock(Settings settings) {
+        super(settings);
     }
 
     @Override
