+++ src/main/java/de/prwh/cobaltmod/core/item/CMAxeItem.java
@@ -1,4 +1,4 @@
-package de.prwh.cobaltmod.core.items;
+package de.prwh.cobaltmod.core.item;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.ImmutableMap;
