+++ src/main/resources/assets/mod_cobalt/blockstates/cobex_chest.json
@@ -1,0 +1,7 @@
+{
+  "variants": {
+    "": {
+      "model": "mod_cobalt:block/chest"
+    }
+  }
+}
