+++ src/main/resources/assets/mod_cobalt/models/block/cobex_chest.json
@@ -1,0 +1,5 @@
+{
+  "textures": {
+    "particle": "mod_cobalt:block/cobex_planks"
+  }
+}
