+++ src/main/resources/assets/mod_cobalt/textures/entity/chest/cobalt_locked.png
