+++ src/main/resources/assets/mod_cobalt/textures/entity/chest/normal.png
