+++ src/main/resources/assets/mod_cobalt/textures/tall_cobex_leaves.png
