+++ src/main/resources/data/mod_cobalt/dimension/cobaldis.json.default
@@ -19,8 +19,8 @@
               0
             ],
             "continentalness": [
-              -0.455,
-              -0.25
+              0,
+              0
             ],
             "erosion": [
               0,
@@ -28,7 +28,7 @@
             ],
             "weirdness": [
               0,
-              0
+              -0
             ],
             "depth": 0,
             "offset": 0
@@ -46,8 +46,8 @@
               0
             ],
             "continentalness": [
-              -0.25,
-              -0.11
+              0,
+              0
             ],
             "erosion": [
               0,
@@ -74,7 +74,7 @@
             ],
             "continentalness": [
               0,
-              0.1
+              0
             ],
             "erosion": [
               0,
@@ -100,8 +100,8 @@
               0
             ],
             "continentalness": [
-              0.1,
-              0.3
+              0,
+              0
             ],
             "erosion": [
               0,
@@ -127,8 +127,8 @@
               0
             ],
             "continentalness": [
-              0.3,
-              0.5
+              0,
+              0
             ],
             "erosion": [
               0,
@@ -154,8 +154,8 @@
               0
             ],
             "continentalness": [
-              0.5,
-              0.75
+              0,
+              0
             ],
             "erosion": [
               0,
@@ -181,8 +181,8 @@
               0
             ],
             "continentalness": [
-              0.75,
-              1
+              0,
+              0
             ],
             "erosion": [
               0,
