+++ src/main/resources/data/mod_cobalt/worldgen/biome/blue_hills.json
@@ -29,8 +29,8 @@
     [],
     [],
     [
-      "mod_cobalt:trees_cobex_forest",
-      "mod_cobalt:patch_grass_normal"
+      "mod_cobalt:patch_grass_normal",
+      "mod_cobalt:trees_cobex_forest"
     ],
     []
   ],
