+++ src/main/resources/data/mod_cobalt/worldgen/biome/cobex_forest.json
@@ -29,8 +29,8 @@
     [],
     [],
     [
-      "mod_cobalt:trees_cobex_forest",
-      "mod_cobalt:patch_grass_forest"
+      "mod_cobalt:patch_grass_forest",
+      "mod_cobalt:trees_cobex_forest"
     ],
     []
   ],
