+++ src/main/resources/data/mod_cobalt/worldgen/biome/flower_plains.json
@@ -29,7 +29,6 @@
     [],
     [],
     [
-      "mod_cobalt:trees_cobex_forest",
       "mod_cobalt:patch_grass_plain"
     ],
     []
