+++ src/main/resources/data/mod_cobalt/worldgen/configured_feature/fancy_cobex_tree.json
