+++ src/main/resources/data/mod_cobalt/worldgen/configured_feature/tall_cobex_tree.json
