+++ src/main/resources/data/mod_cobalt/worldgen/configured_feature/trees_cobex_forest.json
@@ -1,12 +1,16 @@
 {
+  "type": "minecraft:random_selector",
   "config": {
     "features": [
       {
-        "feature": "mod_cobalt:fancy_cobex_tree",
-        "chance": 0.33333334
+        "chance": 0.1,
+        "feature": "mod_cobalt:fancy_cobex_tree"
+      },
+      {
+        "chance": 0.25,
+        "feature": "mod_cobalt:cobex_tree"
       }
     ],
     "default": "mod_cobalt:cobex_tree"
-  },
-  "type": "minecraft:random_selector"
+  }
 }
