+++ src/main/resources/data/mod_cobalt/worldgen/placed_feature/cobex_tree.json
@@ -1,5 +1,5 @@
 {
-  "feature": "mod_cobalt:cobex",
+  "feature": "mod_cobalt:cobex_tree",
   "placement": [
     {
       "predicate": {
