+++ src/main/resources/data/mod_cobalt/worldgen/placed_feature/fancy_cobex_tree.json
@@ -1,5 +1,5 @@
 {
-  "feature": "mod_cobalt:fancy_cobex",
+  "feature": "mod_cobalt:fancy_cobex_tree",
   "placement": [
     {
       "predicate": {
