+++ src/main/resources/data/mod_cobalt/worldgen/placed_feature/tall_cobex_tree.json
@@ -1,5 +1,5 @@
 {
-  "feature": "mod_cobalt:tall_cobex",
+  "feature": "mod_cobalt:tall_cobex_tree",
   "placement": [
     {
       "predicate": {
