+++ src/main/resources/data/mod_cobalt/worldgen/placed_feature/trees_cobex_forest.json
@@ -2,43 +2,39 @@
   "feature": "mod_cobalt:trees_cobex_forest",
   "placement": [
     {
+      "type": "minecraft:count",
       "count": {
+        "type": "minecraft:weighted_list",
         "distribution": [
           {
-            "data": 0,
-            "weight": 19
-          },
-          {
-            "data": 1,
-            "weight": 1
+            "weight": 9,
+            "data": 10
           }
-        ],
-        "type": "minecraft:weighted_list"
-      },
-      "type": "minecraft:count"
+        ]
+      }
     },
     {
       "type": "minecraft:in_square"
     },
     {
-      "max_water_depth": 0,
-      "type": "minecraft:surface_water_depth_filter"
+      "type": "minecraft:surface_water_depth_filter",
+      "max_water_depth": 0
     },
     {
-      "heightmap": "OCEAN_FLOOR",
-      "type": "minecraft:heightmap"
+      "type": "minecraft:heightmap",
+      "heightmap": "OCEAN_FLOOR"
     },
     {
+      "type": "minecraft:block_predicate_filter",
       "predicate": {
+        "type": "minecraft:would_survive",
         "state": {
+          "Name": "mod_cobalt:cobex_sapling",
           "Properties": {
             "stage": "0"
-          },
-          "Name": "mod_cobalt:cobex_sapling"
-        },
-        "type": "minecraft:would_survive"
-      },
-      "type": "minecraft:block_predicate_filter"
+          }
+        }
+      }
     },
     {
       "type": "minecraft:biome"
